upgrade to Node 18 and react-scripts 5 to fix OpenSSL v3 incompatibility
- Updated engines.node from 12.19.0 to 18.x in root and frontend package.json
- Upgraded react-scripts from 3.4.3 to 5.0.1 (webpack 5, OpenSSL v3 compatible)
- Upgraded react/react-dom from 16.13.1 to 17.0.2 (required for jsx-runtime)
- Pinned ajv@^8 to resolve peer dependency conflict in react-scripts 5
- Migrated service-worker.js to Workbox format (CRA 5 standard)
- Extracted SW registration logic to serviceWorkerRegistration.js
- Updated index.js import accordingly
