+++ frontend/package.json
@@ -11,14 +11,15 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
+    "ajv": "^8.18.0",
     "axios": "^0.26.1",
     "history": "^5.0.0",
     "moment": "^2.29.4",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-helmet-async": "^1.0.7",
     "react-moment": "^1.0.0",
-    "react-scripts": "3.4.3",
+    "react-scripts": "5.0.1",
     "react-sweet-state": "^2.3.1",
     "styled-components": "^5.2.0"
   },
@@ -47,6 +48,6 @@
     "http-proxy-middleware": "^1.0.5"
   },
   "engines": {
-    "node": "12.19.0"
+    "node": "18.x"
   }
 }
