+++ package.json
@@ -37,6 +37,6 @@
     "uuid": "^8.3.0"
   },
   "engines": {
-    "node": "12.19.0"
+    "node": "18.x"
   }
 }
